moving some AbstractDynodeRunner code out to utils finishing timeline-> timeseries rename
