+++ examples/config/config_inferer_covid.json
@@ -123,8 +123,8 @@
     "SEASONALITY_SECOND_WAVE": 0.5,
     "SEASONALITY_SHIFT": 0,
     "INFERENCE_PRNGKEY": 8675309,
-    "INFERENCE_NUM_WARMUP": 30,
-    "INFERENCE_NUM_SAMPLES": 30,
+    "INFERENCE_NUM_WARMUP": 150,
+    "INFERENCE_NUM_SAMPLES": 50,
     "INFERENCE_NUM_CHAINS": 4,
     "INFERENCE_PROGRESS_BAR": true,
     "MODEL_RAND_SEED": 8675309
