+++ src/dynode/dynode_runner.py
@@ -6,20 +6,18 @@ aids the user in the production of timeseries to describe a model run
 It also handles the saving of stderr and stdout copies as the job executes.
 """
 
-import copy
 import json
 import os
 import warnings
 from abc import ABC, abstractmethod
-from typing import Optional, Union
+from typing import Union
 
 import numpy as np
 import pandas as pd  # type: ignore
 from diffrax import Solution  # type: ignore
 from jax import Array
 
-from . import SEIC_Compartments, utils, vis_utils
-from .abstract_parameters import AbstractParameters
+from . import utils, vis_utils
 from .mechanistic_inferer import MechanisticInferer
 from .static_value_parameters import StaticValueParameters
 
@@ -80,285 +78,6 @@ class AbstractDynodeRunner(ABC):
         with open(new_config_path, "w") as j:
             # open the location of `config_path` locally, and save it to the new_config_path
             json.dump(json.load(open(config_path, "r")), j, indent=4)
-
-    def match_index_len(
-        self, series: Array, index_len: int, pad: str = "l"
-    ) -> np.ndarray:
-        """A helper function designed to simply insert Nans on the left or right
-        of a series so that it matches the desired `index_len`"""
-
-        def _pad_fn(series, index_len, pad):
-            if "l" in pad:
-                return np.pad(
-                    series,
-                    (index_len - len(series), 0),
-                    "constant",
-                    constant_values=None,
-                )
-            elif "r" in pad:
-                return np.pad(
-                    series,
-                    (0, index_len - len(series)),
-                    "constant",
-                    constant_values=None,
-                )
-
-        if len(series) < index_len:
-            return _pad_fn(series, index_len, pad)
-        return series
-
-    def _get_vaccination_timeseries(
-        self, vaccination_func, num_days_predicted
-    ) -> np.ndarray:
-        """gets num individuals vaccinated by day and age bin
-
-        Parameters
-        ----------
-        vaccination_func : Callable[int]
-            function to generate vaccinations on a given day
-        num_days_predicted : int
-            number of days the simulation was run for
-
-        Returns
-        -------
-        np.ndarray
-            timeseries of shape (num_days_predicted, age) containing the
-            output of vax_function applied on that day summed across all different vaccine stratifications
-        """
-        return np.array(
-            [
-                np.sum(vaccination_func(t), axis=1)
-                for t in range(num_days_predicted)
-            ]
-        )
-
-    def _get_seasonality_timeseries(
-        self, seasonality_func, num_days_predicted: int
-    ) -> np.ndarray:
-        """gets seasonality coefficient by day
-
-        Parameters
-        ----------
-        seasonality_func : Callable[int]
-            function to generate seasonality coefficients in the model
-        num_days_predicted : int
-            number of days the simulation was run for
-
-        Returns
-        -------
-        np.ndarray
-            timeseries of shape (num_days_predicted,) containing the output of seasonality_func applied on that day
-        """
-        return np.array(
-            [seasonality_func(t) for t in range(num_days_predicted)]
-        )
-
-    def _get_external_infection_timeseries(
-        self,
-        external_i_func,
-        num_days_predicted: int,
-        model: AbstractParameters,
-    ) -> np.ndarray:
-        """generates the external_introduction counts by day and strain
-
-        Parameters
-        ----------
-        external_i_func : Callable[int]
-            function to generate externally introduced people
-        num_days_predicted : int
-            number of days the simulation was run for
-        model : AbstractParameters
-            parameters object for enum lookup
-
-        Returns
-        -------
-        np.ndarray
-            timseries of all external introductions of shape (num_days_predicted, model.config.NUM_STRAINS)
-        """
-        # sum across age groups since we do % of each age bin anyways
-        return np.array(
-            [
-                np.sum(
-                    external_i_func(t),
-                    axis=(
-                        model.config.I_AXIS_IDX.age,
-                        model.config.I_AXIS_IDX.hist,
-                        model.config.I_AXIS_IDX.vax,
-                    ),
-                )
-                for t in range(num_days_predicted)
-            ]
-        )
-
-    def _get_sero_proportion_timeseries(
-        self,
-        compartment_timeseries: SEIC_Compartments,
-        population: Array,
-        model: AbstractParameters,
-    ) -> np.ndarray:
-        """given a timeseries of infections, finds individuals who
-        have never been infected, and uses that to calculate the predicted sero-positive
-        proportion among the population by day.
-
-        Parameters
-        ----------
-        infections : SEIC_Compartments
-            tuple of jax arrays containing a timeseries of each compartment's values
-            on a given day.
-        population : jax.Array
-            an array of len(model.config.NUM_AGE_GROUPS) containing the
-            population sizes of each age bin
-        model : AbstractParameters
-            a parameter object containing a config.S_AXIS_IDX enum for lookups and
-            a config.POPULATION array for population counts
-
-        Returns
-        -------
-        np.ndarray
-            array of two dimensions (time, age) matching the first two dimensions
-            of the Susceptible compartment's timeseries
-        """
-        # select timeseries of those with infect_hist 0, sum over all vax/wane tiers
-        never_infected = np.sum(
-            compartment_timeseries[model.config.COMPARTMENT_IDX.S][
-                :, :, 0, :, :
-            ],
-            axis=(model.config.S_AXIS_IDX.vax, model.config.S_AXIS_IDX.wane),
-        )
-        # 1-never_infected is those sero-positive / POPULATION to make proportions
-        sim_sero = 1 - never_infected / population
-        return sim_sero
-
-    def _generate_model_component_timeseries(
-        self,
-        model: AbstractParameters,
-        solution: Solution,
-        hospitalization_preds: Optional[Array] = None,
-    ) -> pd.DataFrame:
-        """
-        a private function which generates a dataframe of different timeseries of interest, optionally including hospitalizations
-
-        Parameters
-        ----------
-        model : AbstractParameters
-            a class that inherits from AbstractParameters and therefore has a get_parameters() method
-        solution : Solution
-            diffrax Solution object as produced by MechanisticRunner.run() command
-        hospitalization_preds : Optional Array
-            models hospitalization predictions, usually the infections
-            matrix with some infection hospitalization ratio applied,
-            must be of shape (time, age_bin)
-
-        NOTE
-        -------------
-        `infections` and `hospitalization_preds` are assumed to begin at
-        model.config.INIT_DATE.
-
-        Returns
-        -------
-        pd.DataFrame
-            a pandas dataframe with a "date" column along with a number of views of the model output
-        """
-        compartment_timeseries = solution.ys
-        num_days_predicted = compartment_timeseries[
-            model.config.COMPARTMENT_IDX.S
-        ].shape[0]
-        sim_dates = [
-            utils.sim_day_to_date(day, model.config.INIT_DATE)
-            for day in range(num_days_predicted)
-        ]
-        df = pd.DataFrame()
-        df["date"] = sim_dates
-        parameters = model.get_parameters()
-        # save a timeseries of shape (num_days_predicted, age_groups)
-        # sum across vax status
-        vaccination_timeseries = self._get_vaccination_timeseries(
-            vaccination_func=parameters["VACCINATION_RATES"],
-            num_days_predicted=num_days_predicted,
-        )
-        # save a seasonality timeseries of shape (num_days_predicted, )
-        df["seasonality_coef"] = self._get_seasonality_timeseries(
-            seasonality_func=parameters["SEASONALITY"],
-            num_days_predicted=num_days_predicted,
-        )
-        # save external introductions timeseries of shape (num_days_predicted, num_strains)
-        external_i_timeseries = self._get_external_infection_timeseries(
-            external_i_func=parameters["EXTERNAL_I"],
-            num_days_predicted=num_days_predicted,
-            model=model,
-        )
-        sim_sero = self._get_sero_proportion_timeseries(
-            compartment_timeseries=compartment_timeseries,
-            population=parameters["POPULATION"],
-            model=model,
-        )
-
-        for age_bin_str in model.config.AGE_GROUP_STRS:
-            age_bin_idx = model.config.AGE_GROUP_IDX[age_bin_str]
-            age_bin_str = age_bin_str.replace("-", "_")
-            # save ground truth and predicted hosp by age
-            if hospitalization_preds is not None:
-                df["pred_hosp_%s" % (age_bin_str)] = self.match_index_len(
-                    hospitalization_preds[:, age_bin_idx], len(df.index)
-                )
-            # save vaccination rate by age
-            df["vaccination_%s" % (age_bin_str)] = vaccination_timeseries[
-                :, age_bin_idx
-            ]
-            # save sero-positive rate by age
-            df["sero_%s" % (age_bin_str)] = sim_sero[:, age_bin_idx]
-        # get total infection incidence
-        (
-            infection_incidence,
-            _,
-        ) = utils.get_timeseries_from_solution_with_command(
-            compartment_timeseries,
-            model.config.COMPARTMENT_IDX,
-            model.config.WANE_IDX,
-            model.config.STRAIN_IDX,
-            "incidence",
-        )
-        # incidence takes a diff, thus reducing length by 1
-        # since we cant measure change in infections at t=0 we just prepend None
-        # plots can start the next day.
-        df["total_infection_incidence"] = self.match_index_len(
-            infection_incidence, len(df.index)
-        )
-        # get strain proportion by strain for each day
-        (
-            strain_proportions,
-            _,
-        ) = utils.get_timeseries_from_solution_with_command(
-            compartment_timeseries,
-            model.config.COMPARTMENT_IDX,
-            model.config.WANE_IDX,
-            model.config.STRAIN_IDX,
-            "strain_prevalence",
-        )
-        # shape (strain, num_days_predicted) summed across age bins
-        population_immunity = np.mean(
-            utils.get_immunity(model, solution), axis=-1
-        )
-        for s_idx, strain_name in enumerate(
-            model.config.STRAIN_IDX._member_names_
-        ):
-            # save out strain specific data for each strain
-            df["%s_strain_proportion" % strain_name] = strain_proportions[
-                s_idx
-            ]
-            df[
-                "%s_external_introductions" % strain_name
-            ] = external_i_timeseries[:, s_idx]
-            df["%s_average_immunity" % strain_name] = population_immunity[
-                s_idx, :
-            ]
-        return df
-
-    def _save_samples(self, samples, save_path):
-        # convert np arrays to lists
-        for param in samples.keys():
-            samples[param] = samples[param].tolist()
-        json.dump(samples, open(save_path, "w"))
 
     def save_mcmc_chains_plot(
         self,
@@ -461,7 +180,7 @@ class AbstractDynodeRunner(ABC):
                 if not any([prefix in name for prefix in exclude_prefixes])
             }
             save_path = os.path.join(self.azure_output_dir, save_filename)
-            self._save_samples(samples, save_path)
+            utils.save_samples(samples, save_path, indent=0)
             # by default save an accompanying mcmc chains plot for readability
             if save_chains_plot:
                 self.save_mcmc_chains_plot(samples)
@@ -522,7 +241,7 @@ class AbstractDynodeRunner(ABC):
             # if it is empty, warn the user, save nothing
             if timesteps:
                 save_path = os.path.join(self.azure_output_dir, save_filename)
-                self._save_samples(timesteps, save_path)
+                utils.save_samples(timesteps, save_path)
             else:
                 warnings.warn(
                     "attempting to call `save_inference_timesteps` but failed to find any timesteps with prefix %s"
@@ -533,164 +252,34 @@ class AbstractDynodeRunner(ABC):
                 "attempting to call `save_inference_timesteps` before inference is complete. Something is likely wrong..."
             )
 
-    def generate_inference_timeseries(
-        self,
-        inferer: MechanisticInferer,
-        particles: list[tuple[int]],
-        save_filename: str = None,
-        extra_timeseries: pd.DataFrame = None,
-        tf: Union[int, None] = None,
-        external_particle: dict[str, Array] = {},
-        verbose: bool = False,
-        max_workers: int = None,
-    ) -> pd.DataFrame:
-        """saves history of inferer sampled values for use by the azure visualizer.
-        saves CSV file to `self.azure_output_dir/save_filename` if `save_filename` is not None.
-        Will error if inferer.infer() has not been run previous to this call
-        except if an external_particle is passed.
-
-        Parameters
-        ----------
-        inferer: MechanisticInferer
-            the inferer object used to sample the parameter chains that will be visualized
-        particles: list[tuple[int, int]]
-            a list of tuples, each of which specifies the (chain_num, particle_num) to load
-            will error if values are out of range of what was sampled.
-        save_filename : str, optional
-            filename to be saved under, if None does not save dataframe and only returns it.
-            DONT CHANGE WITHOUT MODIFICATION to downstream postprocessing scripts, by default None
-        extra_timeseries: pd.DataFrame, optional
-            a pandas dataframe containing a `date` column along with additional columns you wish
-            to be recorded. Dates predicted by `inferer` not included in `extra_timeseries` will
-            be filled with `None`. `extra_timeseriess` are added identically to all `particles_saved`
-        tf: Union[int, None]:
-            number of days to run posterior model for, defaults to same number of days used in fitting
-            if possible.
-        external_posteriors: dict
-            for use of particles defined somewhere outside of this instance of the MechanisticInferer.
-            For example, loading a checkpoint.json containing saved posteriors from an Azure Batch job.
-            expects keys that match those given to `numpyro.sample` often from
-            inference_algo.get_samples(group_by_chain=True).
-        verbose: bool, optional
-            whether or not to pring out the current chain_particle value being executed
-        max_workers: int, optional
-            The maximum number of processes that can be used to
-            load and generate particle timeseries. If None or not given then as many worker
-            processes will be created as the machine has processors. By default, None
-
-        Returns
-        -------
-        pd.DataFrame
-            dataframe containing all particles and their corresponding timeseries.
-        """
-        from concurrent.futures import ProcessPoolExecutor
-
-        if isinstance(extra_timeseries, pd.DataFrame):
-            assert (
-                "date" in extra_timeseries.columns
-            ), "extra_timeseries lacks a `date` column, "
-            "can not be certain of when these observations occur"
-            # attempt conversion to datetime column if it is not already
-            try:
-                extra_timeseries["date"] = pd.to_datetime(
-                    extra_timeseries["date"]
-                )
-            except Exception as e:
-                # we tried to cast `date` column to datetime and it failed, print and reraise error
-                print(
-                    "Encountered an error trying to parse extra_timeseries[date] into a datetime column"
-                )
-                raise e
-
-        class _spoof_static_inferer(MechanisticInferer):
-            """A super simple spoof class made to mimic the current state of an
-            inferer object but without resampling the parameters, instead
-            returning a pre-specified dictionary `static_params`
-            """
-
-            def __init__(self, inferer, static_params):
-                self.__dict__ = inferer.__dict__
-                self.static_params = static_params
-
-            def get_parameters(self):
-                return self.static_params
-
-        all_particles_df = pd.DataFrame()
-
-        def _load_and_generate_particle_timeseries(
-            particle: tuple[int],
-        ) -> pd.DataFrame:
-            # a helper function designed to load a single particle and generate
-            # its timeseries of note, meant to parallelize processes
-
-            # load particle posteriors and their solution dicts
-            posteriors = inferer.load_posterior_particle(
-                particle,
-                tf=tf,
-                external_particle=external_particle,
-                verbose=verbose,
-            )
-            chain_num, particle_num = particle
-            sol_dct = posteriors[particle]
-            # for (chain, particle), sol_dct in posteriors.items():
-            # content of `sol_dct` depends on return value of inferer.run_simulation func
-            infection_timeseries: Solution = sol_dct["solution"]
-            hospitalizations: Array = sol_dct["hospitalizations"]
-            static_parameters: dict[str, Array] = sol_dct["parameters"]
-            # spoof the inferer to return our static parameters when calling `get_parameters()`
-            # instead of trying to sample like it normally does
-            spoof_static_inferer = _spoof_static_inferer(
-                inferer, static_parameters
-            )
-            df = self._generate_model_component_timeseries(
-                spoof_static_inferer,
-                infection_timeseries,
-                hospitalization_preds=hospitalizations,
-            )
-            df["chain_particle"] = "%s_%s" % (chain_num, particle_num)
-            # add user specified extra timeseries, filling in missing dates
-            if isinstance(extra_timeseries, pd.DataFrame):
-                df = df.merge(extra_timeseries, on="date", how="left")
-            return df
-
-        # parallelize the individual particle load/generate process across cpus
-        with ProcessPoolExecutor(max_workers=max_workers) as executor:
-            dfs = executor.map(
-                _load_and_generate_particle_timeseries, particles
-            )
-            # add all chain/particle combo into a main df
-            all_particles_df = pd.concat(list(dfs), ignore_index=True)
-
-        if save_filename:
-            all_particles_df.to_csv(
-                os.path.join(self.azure_output_dir, save_filename), index=False
-            )
-        return all_particles_df
-
     def save_inference_timeseries(
         self,
         inferer: MechanisticInferer,
+        particles: list[tuple[int]],
         timeseries_filename: str = "simulation_timeseries.csv",
-        particles_saved=1,
         extra_timeseries: pd.DataFrame = None,
         tf: Union[int, None] = None,
         external_particle: dict[str, Array] = {},
         verbose: bool = False,
-    ) -> str:
-        """saves history of inferer sampled values for use by the azure visualizer.
-        saves CSV file to `self.azure_output_dir/timeseries_filename`.
-        Look at `shiny_visualizers/azure_visualizer.py` for logic on parsing and visualizing the chains.
-        Will error if inferer.infer() has not been run previous to this call or an external_particle is passed.
+    ) -> pd.DataFrame:
+        """saves model timeseries of important metrics of interest such as
+        vaccination, infections, variant proportions etc for a given
+        simulation period. If the model passed is StaticValueParameters,
+        all inference logic for choosing posterior particles is ignored and a
+        single "particle" is simulated.
 
         Parameters
         ----------
         inferer: MechanisticInferer
             the inferer object used to sample the parameter chains that will be visualized
+        particles: list[tuple[int]]
+            list of (chain, particle) tuples identifying which posteriors you
+            wish to save, values must be less than
+            `model.config.INFERENCE_NUM_CHAINS/SAMPLES` respectively.
         timeseries_filename : str, optional
-            filename to be saved under.
-            DONT CHANGE WITHOUT MODIFICATION to downstream postprocessing scripts, by default "simulation_timeseries.csv"
-        particles_saved : int, optional
-            the number of particles per chain to save timeseries for, by default 1
+            filename to be saved under. None does not save file.
+            DONT CHANGE WITHOUT MODIFICATION to downstream postprocessing
+            scripts, by default "simulation_timeseries.csv"
         extra_timeseries: pd.DataFrame, optional
             a pandas dataframe containing a `date` column along with additional columns you wish
             to be recorded. Dates predicted by `inferer` not included in `extra_timeseries` will
@@ -708,82 +297,63 @@ class AbstractDynodeRunner(ABC):
 
         Returns
         -------
-        str
-            path the inference timeseries were saved to
+        pd.DataFrame
+            dataframe containing all particles and their corresponding timeseries.
+
         """
-        inference_visuals_save_path = os.path.join(
-            self.azure_output_dir,
-            timeseries_filename,
-        )
-        if isinstance(extra_timeseries, pd.DataFrame):
-            assert (
-                "date" in extra_timeseries.columns
-            ), "extra_timeseries lacks a `date` column, "
-            "can not be certain of when these observations occur"
-            # attempt conversion to datetime column if it is not already
-            try:
-                extra_timeseries["date"] = pd.to_datetime(
-                    extra_timeseries["date"]
-                )
-            except Exception as e:
-                # we tried to cast `date` column to datetime and it failed, print and reraise error
-                print(
-                    "Encountered an error trying to parse extra_timeseries[date] into a datetime column"
-                )
-                raise e
+        # Validate inputs before proceeding with main logic.
+        extra_timeseries = self._validate_extra_timeseries(extra_timeseries)
+        self._validate_particles(particles, inferer)
 
         all_particles_df = pd.DataFrame()
-        # randomly select `particles_saved` particles from the number of samples run
-        for particle in np.random.choice(
-            range(inferer.config.INFERENCE_NUM_SAMPLES),
-            particles_saved,
-            replace=False,
-        ):
-            # select this random particle for each of our chains
-            chain_particle_pairs = [
-                (chain, particle)
-                for chain in range(inferer.config.INFERENCE_NUM_CHAINS)
-            ]
-            # assuming the inferer has finished fitting
-            # load those particle posteriors and their solution dicts
-            posteriors = inferer.load_posterior_particle(
-                chain_particle_pairs,
-                tf=tf,
-                external_particle=external_particle,
-                verbose=verbose,
+        # lookup each posterior particle and run simulation with that posterior
+        posteriors = inferer.load_posterior_particle(
+            particles,
+            tf=tf,
+            external_particle=external_particle,
+            verbose=verbose,
+        )
+        for (chain, particle), sol_dct in posteriors.items():
+            # content of `sol_dct` depends on return value of inferer.run_simulation func
+            infection_timeseries: Solution = sol_dct["solution"]
+            hospitalizations: Array = sol_dct["hospitalizations"]
+            static_parameters: dict[str, Array] = sol_dct["parameters"]
+            # spoof the inferer to return our static parameters when calling `get_parameters()`
+            # instead of trying to sample like it normally does
+            spoof_static_inferer = _spoof_static_inferer(
+                inferer, static_parameters
             )
-            for (chain, particle), sol_dct in posteriors.items():
-                # content of `sol_dct` depends on return value of inferer.likelihood func
-                infection_timeseries: Solution = sol_dct["solution"]
-                hospitalizations: Array = sol_dct["hospitalizations"]
-                static_parameters: dict[str, Array] = sol_dct["parameters"]
-                # spoof the inferer to return our static parameters when calling `get_parameters()`
-                # instead of trying to sample like it normally does
-                spoof_static_inferer = copy.copy(inferer)
-                spoof_static_inferer.get_parameters = lambda: static_parameters
-                df = self._generate_model_component_timeseries(
-                    spoof_static_inferer,
-                    infection_timeseries,
-                    hospitalization_preds=hospitalizations,
-                )
-                df["chain_particle"] = "%s_%s" % (chain, particle)
-                # add user specified extra timeseriess, filling in missing dates
-                if isinstance(extra_timeseries, pd.DataFrame):
-                    df = df.merge(extra_timeseries, on="date", how="left")
+            df = utils.generate_model_component_timeseries(
+                spoof_static_inferer,
+                infection_timeseries,
+                hospitalization_preds=hospitalizations,
+            )
+            df["chain_particle"] = "%s_%s" % (chain, particle)
+            # add user specified extra timeseriess, filling in missing dates
+            if isinstance(extra_timeseries, pd.DataFrame):
+                df = df.merge(extra_timeseries, on="date", how="left")
 
-                # add this chain/particle combo onto the main df
-                all_particles_df = pd.concat(
-                    [all_particles_df, df], axis=0, ignore_index=True
-                )
-        all_particles_df.to_csv(inference_visuals_save_path, index=False)
-        return inference_visuals_save_path
+            # add this chain/particle combo onto the main df
+            all_particles_df = pd.concat(
+                [all_particles_df, df], axis=0, ignore_index=True
+            )
+        # if not none, save csv
+        if timeseries_filename:
+            all_particles_df.to_csv(
+                os.path.join(
+                    self.azure_output_dir,
+                    timeseries_filename,
+                ),
+                index=False,
+            )
+        return all_particles_df
 
     def save_static_run_timeseries(
         self,
         parameters: StaticValueParameters,
         sol: Solution,
         timeseries_filename: str = "simulation_timeseries.csv",
-    ):
+    ) -> pd.DataFrame:
         """given a tuple of compartment timeseries, saves a number of timeseries of interest for future visualization
         usually `sol` is retrieved from `diffrax.Solution.ys` which is an object returned from `MechanisticRunner.run()`
 
@@ -801,16 +371,65 @@ class AbstractDynodeRunner(ABC):
 
         Returns
         -------
-        str
-            path the inference timeseries were saved to
+        pd.DataFrame
+            dataframe containing timeseries of metrics of note.
         """
-        inference_visuals_save_path = os.path.join(
-            self.azure_output_dir,
-            timeseries_filename,
-        )
-        # get a number of timeseries to visualize the run
-        df = self._generate_model_component_timeseries(parameters, sol)
+        df = utils.generate_model_component_timeseries(parameters, sol)
         # there is no chain nor particle in a static run, so we save as na_na
         df["chain_particle"] = "na_na"
-        df.to_csv(inference_visuals_save_path, index=False)
-        return inference_visuals_save_path
+        if timeseries_filename:
+            df.to_csv(
+                os.path.join(
+                    self.azure_output_dir,
+                    timeseries_filename,
+                ),
+                index=False,
+            )
+        return df
+
+    def _validate_extra_timeseries(
+        self, extra_timeseries: pd.DataFrame
+    ) -> pd.DataFrame:
+        """Validates the extra timeseries DataFrame."""
+        if isinstance(extra_timeseries, pd.DataFrame):
+            assert (
+                "date" in extra_timeseries.columns
+            ), "extra_timeseries lacks a 'date' column; cannot determine when these observations occur."
+            try:
+                extra_timeseries["date"] = pd.to_datetime(
+                    extra_timeseries["date"]
+                )
+            except Exception as e:
+                print(
+                    "Encountered an error trying to parse extra_timeseries['date'] into a datetime column."
+                )
+                raise e
+        return extra_timeseries
+
+    def _validate_particles(
+        self, particles: list[tuple[int]], inferer: MechanisticInferer
+    ):
+        """Validates the provided particles against configuration limits."""
+        for particle in particles:
+            chain_num, particle_num = particle
+
+            assert (chain_num < inferer.config.INFERENCE_NUM_CHAINS) and (
+                particle_num < inferer.config.INFERENCE_NUM_SAMPLES
+            ), (
+                f"Selected (chain, particle) ({chain_num}, {particle_num}) is out of range "
+                f"of ({inferer.config.INFERENCE_NUM_CHAINS}, {inferer.config.INFERENCE_NUM_SAMPLES})."
+            )
+
+
+class _spoof_static_inferer(MechanisticInferer):
+    """A super simple spoof class made to mimic the current state of an
+    inferer object but without resampling the parameters, instead
+    returning a pre-specified dictionary `static_params`
+    """
+
+    def __init__(self, inferer, static_params):
+        self.__dict__ = inferer.__dict__
+        self.static_params = static_params
+
+    def get_parameters(self):
+        return self.static_params
