+++ src/dynode/utils.py
@@ -8,6 +8,7 @@ import sys
 # importing under a different name because mypy static type hinter
 # strongly dislikes the IntEnum class.
 from enum import EnumMeta as IntEnum
+from typing import Optional
 
 import epiweeks
 import jax.numpy as jnp
@@ -16,10 +17,11 @@ import numpy as np
 import numpyro  # type: ignore
 import numpyro.distributions as Dist  # type: ignore
 import pandas as pd  # type: ignore
+from diffrax import Solution
 from jax import Array
 from scipy.stats import gamma
 
-from . import SEIC_Timeseries
+from . import SEIC_Compartments, SEIC_Timeseries
 
 pd.options.mode.chained_assignment = None
 
@@ -196,9 +198,6 @@ def conditional_knots(t, knots, coefficients):
     indicators = jnp.where(t > knots, t - knots, 0)
     # multiply coefficients by 3 since we taking derivative of cubic spline.
     return jnp.sum(indicators**3 * coefficients, axis=-1)
-
-
-# days of separation between each knot
 
 
 def evaluate_cubic_spline(
@@ -1048,6 +1047,33 @@ def drop_keys_with_substring(dct: dict[str], drop_s: str):
     return dct
 
 
+def match_index_len(
+    series: Array, index_len: int, pad: str = "l"
+) -> np.ndarray:
+    """A helper function designed to simply insert Nans on the left or right
+    of a series so that it matches the desired `index_len`"""
+
+    def _pad_fn(series, index_len, pad):
+        if "l" in pad:
+            return np.pad(
+                series,
+                (index_len - len(series), 0),
+                "constant",
+                constant_values=None,
+            )
+        elif "r" in pad:
+            return np.pad(
+                series,
+                (0, index_len - len(series)),
+                "constant",
+                constant_values=None,
+            )
+
+    if len(series) < index_len:
+        return _pad_fn(series, index_len, pad)
+    return series
+
+
 # @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
 # DEATH CALCULATION CODE
 # @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
@@ -1378,6 +1404,124 @@ def get_timeseries_from_solution_with_command(
     return is_close_v(np.sum(compartment, axis=dimensions_to_sum_over)), label
 
 
+def get_vaccination_timeseries(
+    vaccination_func, num_days_predicted
+) -> np.ndarray:
+    """gets num individuals vaccinated by day and age bin
+
+    Parameters
+    ----------
+    vaccination_func : Callable[int]
+        function to generate vaccinations on a given day
+    num_days_predicted : int
+        number of days the simulation was run for
+
+    Returns
+    -------
+    np.ndarray
+        timeseries of shape (num_days_predicted, age) containing the
+        output of vax_function applied on that day summed across all different vaccine stratifications
+    """
+    return np.array(
+        [
+            np.sum(vaccination_func(t), axis=1)
+            for t in range(num_days_predicted)
+        ]
+    )
+
+
+def get_seasonality_timeseries(
+    seasonality_func, num_days_predicted: int
+) -> np.ndarray:
+    """gets seasonality coefficient by day
+
+    Parameters
+    ----------
+    seasonality_func : Callable[int]
+        function to generate seasonality coefficients in the model
+    num_days_predicted : int
+        number of days the simulation was run for
+
+    Returns
+    -------
+    np.ndarray
+        timeseries of shape (num_days_predicted,) containing the output of seasonality_func applied on that day
+    """
+    return np.array([seasonality_func(t) for t in range(num_days_predicted)])
+
+
+def get_external_infection_timeseries(
+    external_i_func, num_days_predicted: int, model
+) -> np.ndarray:
+    """generates the external_introduction counts by day and strain
+
+    Parameters
+    ----------
+    external_i_func : Callable[int]
+        function to generate externally introduced people
+    num_days_predicted : int
+        number of days the simulation was run for
+    model : AbstractParameters
+        parameters object for enum lookup
+
+    Returns
+    -------
+    np.ndarray
+        timseries of all external introductions of shape (num_days_predicted, model.config.NUM_STRAINS)
+    """
+    # sum across age groups since we do % of each age bin anyways
+    return np.array(
+        [
+            np.sum(
+                external_i_func(t),
+                axis=(
+                    model.config.I_AXIS_IDX.age,
+                    model.config.I_AXIS_IDX.hist,
+                    model.config.I_AXIS_IDX.vax,
+                ),
+            )
+            for t in range(num_days_predicted)
+        ]
+    )
+
+
+def get_sero_proportion_timeseries(
+    compartment_timeseries: SEIC_Compartments,
+    population: Array,
+    model,
+) -> np.ndarray:
+    """given a timeseries of infections, finds individuals who
+    have never been infected, and uses that to calculate the predicted sero-positive
+    proportion among the population by day.
+
+    Parameters
+    ----------
+    infections : SEIC_Compartments
+        tuple of jax arrays containing a timeseries of each compartment's values
+        on a given day.
+    population : jax.Array
+        an array of len(model.config.NUM_AGE_GROUPS) containing the
+        population sizes of each age bin
+    model : AbstractParameters
+        a parameter object containing a config.S_AXIS_IDX enum for lookups and
+        a config.POPULATION array for population counts
+
+    Returns
+    -------
+    np.ndarray
+        array of two dimensions (time, age) matching the first two dimensions
+        of the Susceptible compartment's timeseries
+    """
+    # select timeseries of those with infect_hist 0, sum over all vax/wane tiers
+    never_infected = np.sum(
+        compartment_timeseries[model.config.COMPARTMENT_IDX.S][:, :, 0, :, :],
+        axis=(model.config.S_AXIS_IDX.vax, model.config.S_AXIS_IDX.wane),
+    )
+    # 1-never_infected is those sero-positive / POPULATION to make proportions
+    sim_sero = 1 - never_infected / population
+    return sim_sero
+
+
 def get_var_proportions(inferer, solution):
     """
     Calculate _daily_ variant proportions based on a simulation run.
@@ -1407,37 +1551,6 @@ def get_var_proportions(inferer, solution):
     strain_incidence = jnp.diff(strain_incidence, axis=0)
     sim_vars = strain_incidence / jnp.sum(strain_incidence, axis=-1)[:, None]
     return sim_vars
-
-
-def get_seroprevalence(inferer, solution):
-    """
-    Calculate the seroprevalence (more precisely the cumulative attack rate) based on
-    a simulation run.
-
-    Parameters
-    ----------
-    `inferer` : AbstractParameters
-        an AbstractParameters (e.g., MechanisticInferer or StaticValueParameters) that
-        is used to produce `solution`.
-    `solution`: tuple(jnp.array)
-        solution object that comes out from an ODE run (specifically through
-        `diffrax.diffeqsolve`)
-
-    Returns
-    ----------
-    jnp.array:
-        an array of seroprevalence by the shape of (num_days, NUM_AGE_GROUPS)
-    """
-    never_infected = jnp.sum(
-        solution.ys[inferer.config.COMPARTMENT_IDX.S][:, :, 0, :, :],
-        axis=(
-            # offset for day dimension, un-offset by infection history
-            inferer.config.S_AXIS_IDX.vax,
-            inferer.config.S_AXIS_IDX.wane,
-        ),
-    )
-    sim_sero = 1 - never_infected / inferer.config.POPULATION
-    return sim_sero
 
 
 def get_foi_suscept(p, force_of_infection):
@@ -1535,25 +1648,123 @@ def get_immunity(inferer, solution):
     return immunity_strain
 
 
-def get_vaccination_rates(inferer, num_day):
+def generate_model_component_timeseries(
+    model,
+    solution: Solution,
+    hospitalization_preds: Optional[Array] = None,
+) -> pd.DataFrame:
     """
-    Calculate _daily_ vaccination rates over the course of `num_day`.
+    a private function which generates a dataframe of different timeseries of interest, optionally including hospitalizations
 
     Parameters
     ----------
-    `inferer` : AbstractParameters
-        an AbstractParameters (e.g., MechanisticInferer or StaticValueParameters) that
-        is used to produce `solution`.
-    `num_day`: int
-        number of simulation days
+    model : AbstractParameters
+        a class that inherits from AbstractParameters and therefore has a get_parameters() method
+    solution : Solution
+        diffrax Solution object as produced by MechanisticRunner.run() command
+    hospitalization_preds : Optional Array
+        models hospitalization predictions, usually the infections
+        matrix with some infection hospitalization ratio applied,
+        must be of shape (time, age_bin)
+
+    NOTE
+    -------------
+    `infections` and `hospitalization_preds` are assumed to begin at
+    model.config.INIT_DATE.
 
     Returns
-    ----------
-    list:
-        list of `num_day` vaccination rates arrays, each by the shape of (NUM_AGE_GROUPS,
-        MAX_VACCINATION_COUNT + 1)
+    -------
+    pd.DataFrame
+        a pandas dataframe with a "date" column along with a number of views of the model output
     """
-    return [inferer.vaccination_rate(t).tolist() for t in range(num_day)]
+    compartment_timeseries = solution.ys
+    num_days_predicted = compartment_timeseries[
+        model.config.COMPARTMENT_IDX.S
+    ].shape[0]
+    sim_dates = [
+        sim_day_to_date(day, model.config.INIT_DATE)
+        for day in range(num_days_predicted)
+    ]
+    df = pd.DataFrame()
+    df["date"] = sim_dates
+    parameters = model.get_parameters()
+    # save a timeseries of shape (num_days_predicted, age_groups)
+    # sum across vax status
+    vaccination_timeseries = get_vaccination_timeseries(
+        vaccination_func=parameters["VACCINATION_RATES"],
+        num_days_predicted=num_days_predicted,
+    )
+    # save a seasonality timeseries of shape (num_days_predicted, )
+    df["seasonality_coef"] = get_seasonality_timeseries(
+        seasonality_func=parameters["SEASONALITY"],
+        num_days_predicted=num_days_predicted,
+    )
+    # save external introductions timeseries of shape (num_days_predicted, num_strains)
+    external_i_timeseries = get_external_infection_timeseries(
+        external_i_func=parameters["EXTERNAL_I"],
+        num_days_predicted=num_days_predicted,
+        model=model,
+    )
+    sim_sero = get_sero_proportion_timeseries(
+        compartment_timeseries=compartment_timeseries,
+        population=parameters["POPULATION"],
+        model=model,
+    )
+
+    for age_bin_str in model.config.AGE_GROUP_STRS:
+        age_bin_idx = model.config.AGE_GROUP_IDX[age_bin_str]
+        age_bin_str = age_bin_str.replace("-", "_")
+        # save ground truth and predicted hosp by age
+        if hospitalization_preds is not None:
+            df["pred_hosp_%s" % (age_bin_str)] = match_index_len(
+                hospitalization_preds[:, age_bin_idx], len(df.index)
+            )
+        # save vaccination rate by age
+        df["vaccination_%s" % (age_bin_str)] = vaccination_timeseries[
+            :, age_bin_idx
+        ]
+        # save sero-positive rate by age
+        df["sero_%s" % (age_bin_str)] = sim_sero[:, age_bin_idx]
+    # get total infection incidence
+    (
+        infection_incidence,
+        _,
+    ) = get_timeseries_from_solution_with_command(
+        compartment_timeseries,
+        model.config.COMPARTMENT_IDX,
+        model.config.WANE_IDX,
+        model.config.STRAIN_IDX,
+        "incidence",
+    )
+    # incidence takes a diff, thus reducing length by 1
+    # since we cant measure change in infections at t=0 we just prepend None
+    # plots can start the next day.
+    df["total_infection_incidence"] = match_index_len(
+        infection_incidence, len(df.index)
+    )
+    # get strain proportion by strain for each day
+    (
+        strain_proportions,
+        _,
+    ) = get_timeseries_from_solution_with_command(
+        compartment_timeseries,
+        model.config.COMPARTMENT_IDX,
+        model.config.WANE_IDX,
+        model.config.STRAIN_IDX,
+        "strain_prevalence",
+    )
+    # shape (strain, num_days_predicted) summed across age bins
+    population_immunity = np.mean(get_immunity(model, solution), axis=-1)
+    for s_idx, strain_name in enumerate(
+        model.config.STRAIN_IDX._member_names_
+    ):
+        # save out strain specific data for each strain
+        df["%s_strain_proportion" % strain_name] = strain_proportions[s_idx]
+        df["%s_external_introductions" % strain_name] = external_i_timeseries[
+            :, s_idx
+        ]
+        df["%s_average_immunity" % strain_name] = population_immunity[s_idx, :]
+    return df
 
 
 # @@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@
@@ -1900,3 +2111,27 @@ def find_files(
     postprocess_files = glob.glob(pattern, recursive=recursive)
 
     return [os.path.basename(file) for file in postprocess_files]
+
+
+def save_samples(samples: dict[str, Array], save_path: str, indent=None):
+    """a helper method designed to save dictionaries of jax arrays, which by
+    default are not json serializable. Converts all numpy/jax arrays to
+    python lists, preserves shape.
+
+    Parameters
+    ----------
+    samples : dict[str, Array]
+        dictionary with str keys each containing a jax array of samples or
+        posteriors, often returned by MCMC.get_samples()
+    save_path : str
+        path, relative or absolute to save json to
+    indent : Optional[int]
+        optional indent spaces for pretty-printing of json,
+        None is most compact and least readable, by default None
+    """
+    import json
+
+    # convert np arrays to lists
+    for param in samples.keys():
+        samples[param] = samples[param].tolist()
+    json.dump(samples, open(save_path, "w"), indent=indent)
